+++ lib/flutter_essentials_kit.dart
@@ -1,7 +1,7 @@
 library flutter_essentials_kit;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+// Widgets
+export 'widgets/courtesy_widget.dart';
+
+// Extensions
+export 'extensions/iterable.dart';