+++ test/test_utils/test_widget.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class TestWidget extends StatelessWidget {
+  final Widget child;
+
+  TestWidget({Key key, @required this.child}) : super(key: key) {
+    assert(child != null);
+  }
+
+  @override
+  Widget build(BuildContext context) => MaterialApp(
+        home: child,
+      );
+}