+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## [1.0.0] - TODO: Add release date.
-
-* TODO: Describe initial release.
+## 1.0.0
+* Create CourtesyWidget.
+* Add `filterNull()` on **Iterable** extension.